+++ src/modules/mod.rs
@@ -1,2 +1,3 @@
+pub mod invoice;
 pub mod merchant;
 pub mod wallet;